Axel/add mail (#13)
* Event mail order created

* added update order status event

* added deleted order event

+++ src/eshop.services/ordering/Ordering.Application/Features/Orders/Commands/DeleteOrder/DeleteOrderCommandHandler.cs
@@ -1,4 +1,6 @@
 using BuildingBlocks.CQRS;
+using BuildingBlocks.Messaging.Events;
+using MassTransit;
 using Microsoft.Extensions.Logging;
 using Ordering.Application.Features.Orders.Data;
 using Ordering.Domain.Exceptions;
@@ -8,7 +10,12 @@ namespace Ordering.Application.Features.Orders.Commands.DeleteOrder;
 /// <summary>
 /// Handles the deletion of orders from the system.
 /// </summary>
-public class DeleteOrderCommandHandler(IOrderRepository orderRepository, ILogger<DeleteOrderCommandHandler> logger) 
+public class DeleteOrderCommandHandler(
+    IOrderRepository orderRepository,
+    ILogger<DeleteOrderCommandHandler> logger,
+    IPublishEndpoint publishEndpoint
+)
+     
     : ICommandHandler<DeleteOrderCommand, DeleteOrderCommandResult>
 {
     /// <summary>
@@ -24,6 +31,14 @@ public class DeleteOrderCommandHandler(IOrderRepository orderRepository, ILogger
     {
         logger.LogInformation("Deleting order {OrderId}", request.OrderId);
 
+        var order = await orderRepository.GetByIdAsync(request.OrderId, cancellationToken);
+        if (order == null)
+        {
+            var message = $"Order with ID {request.OrderId} not found";
+            logger.LogWarning(message);
+            throw new OrderNotFoundException(message);
+        }
+
         var success = await orderRepository.DeleteAsync(request.OrderId, cancellationToken);
         if (!success)
         {
@@ -31,6 +46,29 @@ public class DeleteOrderCommandHandler(IOrderRepository orderRepository, ILogger
             logger.LogWarning(message);
             throw new OrderNotFoundException(request.OrderId);
         }
+
+        var orderDeletedEvent = new OrderDeletedEvent
+        {
+            OrderId = order.Id.Value,
+            CustomerId = order.CustomerId.Value,
+            OrderName = order.OrderName.Value,
+            OrderStatus = order.OrderStatus.ToString(),
+            PreviousStatus = "Deleted",
+            TotalPrice = order.TotalPrice,
+            UpdatedDate = DateTime.UtcNow,
+            CustomerName = $"{order.ShippingAddress.FirstName} {order.ShippingAddress.LastName}",
+            CustomerEmail = order.ShippingAddress.EmailAddress,
+            OrderItems = order.OrderItems.Select(oi => new OrderItemEvent
+            {
+                ProductId = oi.ProductId.Value,
+                ProductName = $"Produit #{oi.ProductId.Value}",
+                Quantity = oi.Quantity,
+                Price = oi.Price
+            }).ToList()
+
+        };
+
+        await publishEndpoint.Publish(orderDeletedEvent, cancellationToken);
         
         logger.LogInformation("Order {OrderId} deleted successfully", request.OrderId);
